V3.76 (#419)
* bumping SEMVER to v3.76.1

* codacy code hygiene (#408)


* WIP: redesigning the API of posit randoms to be consistent with exhaustive regression tests

* unifying the type interface to posit regression tests across exhaustive and random

* unifying regression test APIs for conversion

* bug fix in posit conversion regression test

* code hygiene for clang

* removing superfluous test and configuration feedback

* WIP: adding new number system takum: api test and readme

* WIP: regression suite placeholders

* removing deprecated skeleton structure

* adding a two-parameter skeleton

* clean up of 2-parameter skeleton

* WIP: takum skeleton

* WIP: compiling the takum skeleton

* adding a new constexpr test to lns

* making takum constexpr

* code hygiene for clang/gcc

* code hygiene for clang/gcc

* disabling constexpr for clang and gcc

* WIP: value conversion

* WIP: removing second parameter from takum specification

* WIP: conversion to native floating-point, adding to_binary()

* WIP: adding color_print for takums

* WIP: bug fix in to_binary() and color_print()

* WIP: takum value generation

* restructuring the FDP structure to provide a compilation guarded overload for BLAS

* Standardizing the posit source code headers to include a SPDX-License-Identifier

* Standardizing the lns source code headers to include a SPDX-License-Identifier

* Standardizing the native and verification source code headers to include a SPDX-License-Identifier

* Standardizing the takum source code headers to include a SPDX-License-Identifier

* WIP: restructuring the overload of custom BLAS algorithms

* restructuring default BLAS and adding an extension and a modification capability for custom arithmetic

* WIP: LU iterative refinement

* WIP: LU iterative refinement algorith RoundAndReplace

* WIP: iterative refinement RoundAndReplace for set of IEEE floating-points

* WIP: monster check-in for new IR infrastructure and file I/O based test matrices

* WIP: small test matrix data file

* WIP: IR algorithm decoupling

* WIP: data collection serialization

* WIP: datafile serialization name change of test

* WIP: new test matrices data dump

* converted static test matrix include files to an API that uses files with the matrix data

* bug fix of dbns arithmetic that had slipped in due to incorrect bit pattern generation for NaN encoding

* WIP: generating experiment tables

* found bug in cfloat to cfloat conversions, replacing with marshalling through double

* WIP: round and replace experiment including try/catch blocks to deal with numerical exceptions

* Full Round-and-Replace experiment logic

* sota structure for IR experimentation

* WIP: adding Scale-and-Round preconditioner experiment to LUIR algorithmic testing

* adding two-sided Scale and Round preconditioner to the LUIR test suite

* simple restructuring to make it easier to create result files

* Adding non-fused posit results to LUIR experiment capability

* Consistent structure for LUIR experiments with and without fused BLAS operators

* replacing adaptive u_W with absolute 1.0e-6 error cut-off

* update cfloat headers to include SPDX-License

* SFINAE manipulators for cfloat

* SFINAE manipulators for takum

* adding test suite skeletons for conversions to takums

* API change: fraction(number) is now a free function to obtain the fraction of a number

* WIP: linear takum implementation

* compilation fix

* header update on fixpnt

* WIP: RCAing clang complex<fixpnt> failures

* clang code hygiene: MacOS failure is Mac SDK related

* Turning MacOS build off till we figure out what broke on MacOS SDK with complex<> and the standard library

---------

+++ include/universal/native/integers.hpp
@@ -1,7 +1,8 @@
 #pragma once
 // integers.hpp: manipulators for native integer types
 //
-// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017 Stillwater Supercomputing, Inc.
+// SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <cstdint>
@@ -172,7 +173,7 @@ inline std::string to_hex(const Integer& number, bool nibbleMarker = false, bool
 	uint64_t nibble{ 0 };
 	unsigned nibbleIndex = nbits / 4 - 1u;
 	unsigned rightShift = nibbleIndex * 4;
-	for (int i = nbits - 1; i >= 0; --i) {
+	for (int i = static_cast<int>(nbits - 1u); i >= 0; --i) {
 		nibble |= (mask & number);
 		mask >>= 1;
 		if ((i % 4) == 0) {
